Fix Notification loop not work

+++ utils/auth.py
@@ -8,11 +8,16 @@ import requests
 # reference by https://github.com/iancdev
 
 class Auth:
-    def __init__(self, username, password):
+    def __init__(self, username, password, message=None, interaction=None, bot=None):
         self.username = username
         self.password = password
+        
+        #for 2 factor authentication
+        self.message = message
+        self.interaction = interaction
+        self.bot = bot
 
-    async def authenticate(self, message, bot, interaction):
+    async def authenticate(self):
         
         auth_error = 'An unknown error occurred, sorry'  
         
@@ -38,22 +43,23 @@ class Auth:
             if r.json()['type'] == 'auth':
                 auth_error = 'Your username or password may be incorrect!'
             elif r.json()['type'] == 'multifactor':
-                # 2fa error
-                auth_error = '2FA verify code may be incorrect!'
-                await message.edit('**Enter the 2FA verify code**')
+                if self.message is not None and self.bot is not None and self.interaction is not None:
+                    # 2fa error
+                    auth_error = '2FA verify code may be incorrect!'
+                    await self.message.edit('**Enter the 2FA verify code**')
 
-                try:
-                    respond_message = await bot.wait_for("message", check=lambda msg: msg.author == interaction.user and msg.channel == interaction.channel, timeout=90)
-                except asyncio.TimeoutError:
-                    auth_error = '2 factor authentication is Timeout.'
-                data = {
-                    "type": "multifactor",
-                    "code": respond_message.content,
-                    "rememberDevice": False
-                }
-                await message.edit('\u200B')
-                await respond_message.delete()
-                r = session.put('https://auth.riotgames.com/api/v1/authorization', json=data)
+                    try:
+                        respond_message = await self.bot.wait_for("message", check=lambda msg: msg.author == self.interaction.user and msg.channel == self.interaction.channel, timeout=90)
+                    except asyncio.TimeoutError:
+                        auth_error = '2 factor authentication is Timeout.'
+                    data = {
+                        "type": "multifactor",
+                        "code": respond_message.content,
+                        "rememberDevice": False
+                    }
+                    await self.message.edit('\u200B')
+                    await respond_message.delete()
+                    r = session.put('https://auth.riotgames.com/api/v1/authorization', json=data)
             
             pattern = re.compile('access_token=((?:[a-zA-Z]|\d|\.|-|_)*).*id_token=((?:[a-zA-Z]|\d|\.|-|_)*).*expires_in=(\d*)')
             data = pattern.findall(r.json()['response']['parameters']['uri'])[0]
